# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ActivatePolicyRequest": {
        "properties": {
          "approved_evaluation_id": {
            "maxLength": 128,
            "minLength": 3,
            "title": "Approved Evaluation Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "approved_evaluation_id"
        ],
        "title": "ActivatePolicyRequest",
        "type": "object"
      },
      "ActivatePromptRequest": {
        "properties": {
          "approved_evaluation_id": {
            "maxLength": 128,
            "minLength": 3,
            "title": "Approved Evaluation Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "approved_evaluation_id"
        ],
        "title": "ActivatePromptRequest",
        "type": "object"
      },
      "ActivationResponse": {
        "properties": {
          "approved_evaluation_id": {
            "title": "Approved Evaluation Id",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "status",
          "approved_evaluation_id"
        ],
        "title": "ActivationResponse",
        "type": "object"
      },
      "AnswerCitationItem": {
        "properties": {
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "document_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          }
        },
        "required": [
          "citation_id"
        ],
        "title": "AnswerCitationItem",
        "type": "object"
      },
      "AnswerClaimItem": {
        "properties": {
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Citation Ids",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/AnswerCitationItem"
            },
            "title": "Citations",
            "type": "array"
          },
          "claim_text": {
            "title": "Claim Text",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "support_status": {
            "title": "Support Status",
            "type": "string"
          }
        },
        "required": [
          "ordinal",
          "claim_text",
          "support_status",
          "citation_ids"
        ],
        "title": "AnswerClaimItem",
        "type": "object"
      },
      "AnswerVersionLinks": {
        "properties": {
          "governance": {
            "title": "Governance",
            "type": "string"
          },
          "research_search": {
            "title": "Research Search",
            "type": "string"
          },
          "review_queue": {
            "title": "Review Queue",
            "type": "string"
          },
          "self": {
            "title": "Self",
            "type": "string"
          }
        },
        "required": [
          "self",
          "research_search",
          "review_queue",
          "governance"
        ],
        "title": "AnswerVersionLinks",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchAnswerResponse": {
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_summary": {
            "title": "Answer Summary",
            "type": "string"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AnswerClaimItem"
            },
            "title": "Claims",
            "type": "array"
          },
          "contradictions": {
            "items": {
              "type": "string"
            },
            "title": "Contradictions",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "evidence_context_sha256": {
            "title": "Evidence Context Sha256",
            "type": "string"
          },
          "index_manifest_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Manifest Checksum"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "links": {
            "$ref": "#/components/schemas/AnswerVersionLinks"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_provider_id": {
            "title": "Model Provider Id",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "research_run_id": {
            "title": "Research Run Id",
            "type": "string"
          },
          "retrieval_index_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Index Id"
          },
          "retrieval_run_id": {
            "title": "Retrieval Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uncertainty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertainty"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "answer_id",
          "research_run_id",
          "retrieval_run_id",
          "version",
          "status",
          "answer_summary",
          "model_provider_id",
          "model_id",
          "prompt_version",
          "policy_version",
          "evidence_context_sha256",
          "links"
        ],
        "title": "ResearchAnswerResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "VYU API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/model-gateway/policies/{policy_id}/activate": {
      "post": {
        "operationId": "activate_model_policy_v1_admin_model_gateway_policies__policy_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivatePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Model Policy",
        "tags": [
          "model-gateway-admin"
        ]
      }
    },
...
    "/v1/admin/model-gateway/prompts/{template_id}/activate": {
      "post": {
        "operationId": "activate_prompt_template_v1_admin_model_gateway_prompts__template_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivatePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Prompt Template",
        "tags": [
          "model-gateway-admin"
        ]
      }
    },
...
    "/v1/research/searches/{search_id}/answer": {
      "get": {
        "operationId": "get_research_search_answer_v1_research_searches__search_id__answer_get",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research Search Answer",
        "tags": [
          "research"
        ]
      }
    },
...
  }
}